+++ js_libraries/types/types/common/element/markdown.d.ts
@@ -45,6 +45,13 @@ export interface MarkdownImageTapEvent {
   url: string;
 }
 
+export interface MarkdownTextClickEvent {
+  /**
+   * ID of the clicked text attachment.
+   */
+  id: string;
+}
+
 export interface MarkdownParseEndEvent {
   /**
    * The id of content that parsed.
@@ -233,6 +240,11 @@ export interface MarkdownProps extends StandardProps {
    * @since 3.1
    */
   bindimageTap?: (e: BaseEvent<'bindimageTap', MarkdownImageTapEvent>) => void;
+
+  /**
+   * Callback when a text attachment is clicked.
+   */
+  bindtextClick?: (e: BaseEvent<'bindtextClick', MarkdownTextClickEvent>) => void;
 
   /**
    * Callback when parse ends.